02 RAG For Apple & Tesla/pyproject.toml
[project]
name = "rag-4-finance"
version = "0.1.0"
description = "RAG + LLM system with PDF ingestion, embeddings, reranking, and evaluation"
requires-python = ">=3.9"

dependencies = [
    # Core Dependencies
    "python-dotenv==1.0.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",

    # PDF Processing
    "pypdf==3.17.0",
    "pdfplumber==0.10.3",

    # Re-ranker
    "rank-bm25==0.2.2",
    "flashrank==0.2.3",


    # Embeddings & Vector Store
    "sentence-transformers==2.2.2",
    "faiss-cpu==1.7.4",
    "chromadb==0.4.22",

    # LLM
    "transformers==4.36.0",
    "torch==2.1.0",
    "accelerate==0.25.0",
    "bitsandbytes==0.41.3",

    # Reranking
    "rank-bm25==0.2.2",

    # Utilities
    "numpy==1.24.3",
    "pandas==2.1.4",
    "tqdm==4.66.1",
    "loguru==0.7.2",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",

    # Jupyter
    "jupyter==1.0.0",
    "ipykernel==6.27.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
